requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.2
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.2.3
numpy==1.26.4
matplotlib==3.8.4
seaborn==0.13.2
scipy==1.13.1
duckdb==1.1.2
python-multipart==0.0.6
Pillow==10.2.0
lxml==5.2.2
plotly==5.22.0
scikit-learn==1.5.2
python-dotenv==1.0.1
PyPDF2==3.0.1
pdfplumber==0.11.4
opencv-python==4.9.0.80
pytesseract==0.3.13
SpeechRecognition==3.10.4
pydub==0.25.1
transformers==4.36.2
torch==2.4.1+cpu
aiofiles==23.2.1
setuptools==69.5.1
--extra-index-url https://download.pytorch.org/whl/cpu
